create component thumb img
