+++ src/components/Board/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Container } from "./styles";
 
-export default function Products() {
+export default function Board() {
   return (
     <Container>
       <h1>Products</h1>
